refactor(Dropdown): Make indicator icon reusable
